// types/OrderSteps.ts
export enum OrderSteps {
	WAITING_CONFIRMATION = "WAITING_CONFIRMATION",
	UPDATING_LAST_DATA = "UPDATING_LAST_DATA",
	OPERATION_COMPLETED = "OPERATION_COMPLETED",
}

export enum OrderStatusSteps {
	PLACING_ORDER = "PLACING_ORDER",
	ORDER_PLACED = "ORDER_PLACED",
	ORDER_SHIPPED = "ORDER_SHIPPED",
	ORDER_RECEIVED = "ORDER_RECEIVED",
	ERROR = "ERROR",
}